Fix test suite for Windows
While loading the yar extension via the `$cmd_args` argument of
`yar_server_start()` would work for running the tests via `nmake test`,
that would require in-tree extension builds.  At least for AppVeyor,
the builds are done with phpsize for performance reasons, and as such
we have to set `TEST_PHP_ARGS` anyway, which would then cause warnings
regarding multiple loading of the extension.  Therefore we do not use
any default command line arguments.

Furthermore, we must not connect stderr of the child processes to
stderr of the parent, because that would yield spurious output, which
breaks the test cases.

Because eleven test cases are hanging on Windows for some yet unknown
reason, and since dynamic XFAIL is only available as of PHP 7.4.0, we
skip these tests for now.